Update tests, add import for auth backends.
Auth backends now use the path as the ID, not the name. This will allow
us to have multiple of the same auth backend mounted on the same vault
server. It will also allow us to match the auth backend using just the
path, instead of the path and the type combination.

This is largely necessary for import, where we only have the ID, so we
need the ID to be able to uniquely identify the resource, which the type
can't do, but the path can.

We also updated the initial test to randomise the path the auth backend
is mounted at, instead of mounting it at a static path, which makes
tests more resilient and less likely to fail due to collisions.

Read now sets all the fields in state.