Implement AppRole roles.
Add a new vault_approle_auth_backend_role resource that enables users to
provision and manage AppRole roles.

This also updates to use Terraform 0.10.1 as the vendor, as we need it
for GetOkExists support. But updating to 0.10.8 requires a lot of new
dependencies and was getting us into dependency headaches land, so this
PR is the smallest possible step to achieve what we want.

A future AppRole Secret ID resource is planned, so no support for it is
in this PR. Also, an AppRole RoleID data source is forthcoming, as well,
but it was included here because users can specify their own RoleIDs and
we want to be able to support that using Terraform.