Check for TF_ACC before retrieving metadata info.
Even though our tests that rely on metadata info won't run if TF_ACC
isn't set, they'll still try to retrieve metadata info. Because of the
short timeout for unit tests, this causes the tests to fail.

We also get a weird error about the endpoint not being found, even
though the Available method should gate that, but that's a matter for a
different day.