Fix tests, remove redundant sets.
There were a couple of calls to d.Set that passed in only the
intersection of Groups and Users for a backend. These have been removed,
in favor of just relying on the Read method to set that data in the
backend based on what the API returns.

There was a test that was failing because it was expecting a slice with
an empty string in it, and it was getting an empty slice. I corrected it
to expect an empty slice, and now the tests pass.

I also noticed that if the test was expecting an empty slice, it
wouldn't fail if a slice with items in it was returned. I added a check
to ensure that would cause a failure.