Fix aws auth backend login tests and nonce.
Fix the handling of the aws_auth_backend_login nonce field, as it's only
returned on the first request, and we need to not overwrite it with an
empty string on subsequent reads.

Also, fix the tests so they can be run using an environment variable,
and stop using removed fields in the tests.