Update rabbitmq_secret_backend_role vhost to block.
The vault API accepts the vhost for rabbitmq_secret_backend_role as a
string, but always returns it as a JSON object. There's no Terraform
equivalent (maps can't have complex types as the values), so instead we
turn it into a list (sets were having funky behavior with the hashes,
and I didn't have time to fully look into it). This is a better UX than
a JSON object marshalled into a string in the config, and allows reads
to be consistent. It is, however, a breaking change.