Make affected blocks attributes.
We have some computed optional blocks that will exhibit the computed
optional problem of 0.12. In 0.11, these blocks could be set like so
`block = []` to set the block to an explicit empty value. This could be
used to signal a return to the API default, or as a way to delete the
entry the user entered. Because 0.12 makes a strict delineation between
blocks and attributes, we need to tell Terraform to treat these blocks
as attributes to preserve this behavior. See
https://www.terraform.io/docs/extend/terraform-0.12-compatibility.html#computed-resource-attributes
for more info.